move user input to InputComponent
- replaced deprecated library 'expo-av' with new library 'expo-audio'
- wrapped main view with KeyboardAvoidingView to move view automatically up if the software keyboard is opened
- moved text input, buttons and recorder logic to InputComponent
