+++ package.json
@@ -9,7 +9,9 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@shopify/react-native-skia": "^2.0.5",
     "@siteed/expo-audio-studio": "^2.14.3",
+    "@siteed/expo-audio-ui": "^0.6.0",
     "base64-arraybuffer": "^1.0.2",
     "base64-js": "^1.5.1",
     "expo": "~53.0.10",
@@ -21,6 +23,8 @@
     "morse-pro": "^3.0.1",
     "react": "19.0.0",
     "react-native": "0.79.3",
+    "react-native-gesture-handler": "^2.26.0",
+    "react-native-reanimated": "^3.18.0",
     "wav-file-encoder": "^1.0.4"
   },
   "devDependencies": {
